+++ decidim-core/app/events/decidim/amendable/emendation_promoted_event.rb
@@ -1,7 +1,7 @@
 # frozen-string_literal: true
 
 module Decidim::Amendable
-  class AmendmentAcceptedEvent < Decidim::Amendable::AmendmentBaseEvent
+  class EmendationPromotedEvent < Decidim::Amendable::AmendmentBaseEvent
     i18n_attributes :amendable_path, :amendable_type, :amendable_title, :emendation_path, :emendation_author_nickname, :emendation_author_path
 
     def amendment_resource